+++ leafmap/kepler.py
@@ -469,11 +469,13 @@ class Map(keplergl.KeplerGl):
                 with open(outfile) as f:
                     lines = f.readlines()
                     out_html = "".join(lines)
-                os.remove(outfile)
                 return out_html
 
         except Exception as e:
             raise Exception(e)
+        finally:
+            if not save and os.path.exists(outfile):
+                os.remove(outfile)
 
     def to_streamlit(
         self,
